Refactored Code
Added more distinct packages to show the difference in names
