+++ src/main/java/duke/MainWindow.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.main.Duke;
+import duke.ui.Ui;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
