+++ src/main/java/duke/TaskList.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.ui.Ui;
+
 import java.util.List;
 import java.util.ArrayList;
