+++ src/main/java/duke/events/Deadline.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.events;
 
 /**
  * Deadline is an extension of the Task class
@@ -6,9 +6,11 @@ package duke;
  * And prints it out in a readable format via the toString method
  */
 
+import duke.Task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
+
 public class Deadline extends Task {
     protected LocalDate by;
     private String unformattedDate;
