+++ src/main/java/duke/events/Event.java
@@ -1,8 +1,9 @@
-package duke;
+package duke.events;
+
+import duke.Task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 
 public class Event extends Task {
