+++ src/main/java/duke/exceptions/EmptyTaskException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exceptions;
 public class EmptyTaskException extends Exception {
     public EmptyTaskException(String message) {
         super(message);
