+++ src/main/java/duke/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exceptions;
 public class InvalidCommandException extends Exception {
     public InvalidCommandException(String message) {
         super(message);
