+++ src/main/java/duke/main/Launcher.java
@@ -1,5 +1,4 @@
-package duke;
-
+package duke.main;
 import javafx.application.Application;
 
 /**
