+++ src/main/java/duke/ui/Ui.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.ui;
+
+import duke.Task;
 
 import java.util.List;
