+++ src/test/java/DukeTest.java
@@ -1,4 +1,5 @@
 import duke.*;
+import duke.parser.Parser;
 import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.List;
